Update docs for Qubes 4.1
Made sure to test in dom0 on 4.1. The "disablerepo" flag was necessary
on my recently-migrated machine. Might be overkill, but including for
clarity since it won't do any harm.

The madness about /usr/{local,}/bin/hexagon path in the rpm seems
to be related to whether a virtualenv was used for building. Mea culpa.
On Debian 11 Bullseye, with NO virtualenv, the /usr/local/bin/ path is
used. I can't seem to find good documentation on this behavior,
so I'm letting it flap.